+++ GTpo/src/gtpoGenGraph.h
@@ -406,7 +406,9 @@ public:
 
         SharedNode subGroupNode = std::static_pointer_cast<GenNode<Config>>(subGroup);
         groupNode( weakGroup, WeakNode{subGroupNode} );
-        group->notifyGroupInserted( WeakGroup{subGroup} );
+
+        // FIXME groups
+        //group->notifyGroupInserted( WeakGroup{subGroup} );
     }
 
     /*! \brief Insert an existing node \c weakNode in group \c weakGroup group.
@@ -441,7 +443,9 @@ public:
 
         SharedNode subGroupNode = std::static_pointer_cast<GenNode<Config>>(subGroup);
         ungroupNode( weakGroup, WeakNode{subGroupNode} );
-        group->notifyGroupRemoved( WeakGroup{subGroup} );
+
+        // FIXME groups
+        //group->notifyGroupRemoved( WeakGroup{subGroup} );
     }
 
 private: